Restructures to clearly delineate where code is intended to run. Properly stringifies arguments in monkeypatched fluid.log impl.
