+++ package.json
@@ -12,7 +12,7 @@
     "dependencies": {
         "infusion": "2.0.0-dev.20151007T134312Z.8c33cd4",
         "dedupe-infusion": "1.0.0",
-        "electron-prebuilt": "0.33.7"
+        "electron-prebuilt": "0.34.0"
     },
     "devDependencies": {
         "grunt-contrib-jshint": "0.9.0",
