+++ src/main-thread/window.js
